this was the final code from the competition :more
This code was my final version at the end of the competition, there were several glitches with the controller and the firmware, as well as lots of different things we tried with it (changes in the pit and hotel room) that I just never committed. Oh well. This code could use improvement but is manageable. Onward!